refactor: rename missed custom hook filenames
