+++ apps/calendar/src/components/popup/eventStateSelector.tsx
@@ -4,7 +4,7 @@ import { PopupSection } from '@src/components/popup/popupSection';
 import { StateDropdownMenu } from '@src/components/popup/stateDropdownMenu';
 import { cls } from '@src/helpers/css';
 import { useDropdownState } from '@src/hooks/common/useDropdownState';
-import { FormStateActionType, FormStateDispatcher } from '@src/hooks/popup/formState';
+import { FormStateActionType, FormStateDispatcher } from '@src/hooks/popup/useFormState';
 
 import { EventState } from '@t/events';
