+++ apps/calendar/src/components/timeGrid/movingEventShadow.tsx
@@ -1,7 +1,7 @@
 import { h } from 'preact';
 
 import { TimeEvent } from '@src/components/events/timeEvent';
-import { useTimeGridEventMove } from '@src/hooks/timeGrid/timeGridEventMove';
+import { useTimeGridEventMove } from '@src/hooks/timeGrid/useTimeGridEventMove';
 import { isNil } from '@src/utils/type';
 
 import { GridPositionFinder, TimeGridData } from '@t/grid';
